+++ src/inner-slider.jsx
@@ -1,17 +1,17 @@
 'use strict';
 
 import React from 'react';
-// import EventHandlersMixin from './mixins/event-handlers';
+import EventHandlersMixin from './mixins/event-handlers';
 import HelpersMixin from './mixins/helpers';
 import initialState from './initial-state';
 import defaultProps from './default-props';
 import classnames from 'classnames';
 
 import Track from './track';
-// import Dots from './dots';
+import Dots from './dots';
 
 var Slider = React.createClass({
-  mixins: [HelpersMixin],
+  mixins: [HelpersMixin, EventHandlersMixin],
   getInitialState: function () {
     return initialState;
   },
@@ -35,23 +35,37 @@ var Slider = React.createClass({
   },
   render: function () {
     var className = classnames('slick-initialized', 'slick-slider', this.props.className);
+
+    var trackProps = {
+      infinite: this.props.infinite,
+      centerMode: this.props.centerMode,
+      currentSlide: this.state.currentSlide,
+      slideWidth: this.state.slideWidth,
+      slidesToShow: this.props.slidesToShow,
+      trackStyle: this.state.trackStyle,
+      variableWidth: this.props.variableWidth
+    };
+
+    var dotProps = {
+      dots: this.props.dots,
+      dotsClass: this.props.dotsClass,
+      slideCount: this.state.slideCount,
+      slidesToShow: this.props.slidesToShow,
+      currentSlide: this.state.currentSlide,
+      slidesToScroll: this.props.slidesToScroll,
+      clickHandler: this.changeSlide
+    };
+
     return (
-      <div className={className} >
+      <div className={className}>
         <div
           ref='list'
           className="slick-list">
-        <Track
-          ref='track'
-          infinite={this.props.infinite}
-          centerMode={this.props.centerMode}
-          currentSlide={this.state.currentSlide}
-          slideWidth={this.state.slideWidth}
-          slidesToShow={this.props.slidesToShow}
-          trackStyle={this.state.trackStyle}
-          variableWidth={this.props.variableWidth}>
-          {this.props.children}
-        </Track>
+          <Track ref='track' {...trackProps}>
+            {this.props.children}
+          </Track>
         </div>
+        <Dots {...dotProps}/>
       </div>
     );
   }
